Kafka-pyspark-streaming/requirements.txt
pyspark-3.5.0